requirements.txt
Django >= 4.11
psycopg2-binary >= 2.8.6
redis
celery
djangorestframework
djangorestframework-simplejwt